Remove mod of CMAKE_INSTALL_PREFIX, fix code formatting (PHI Kanban #3232 and #3217)
I removed the speical logic to modify CMAKE_INSTALL_PREFIX.  It was being
hard-coded to append tpls/opt/ or tpls/dbg/.  We don't want to hard-code these
because we want to be able to install more TPL versions like
tpls/dbg-checkedstl/, etc.

I also reduced the indentation a little and made to fit in appox 80 columns to
make more readable.