Change HDF5 version to 1.8.9, make TPL_LIST cache and sorted (PHI Kanban #3232)
There were two different version numbers for the various libs listed in the
CMakeLists.txt.  Not clear which one was being used.  Still not but I
elimiated the duplication.

I made TPL_LIST a proper cache var and I made it so if the user selects the
list, the order they give does not matter.