Add explicit install command, fix for shared libs (PHI Kanban #3232)
The main thing I did here was to change the file extensions from *.a to *.so
when shared libs are built.  They were hard-coded to static libs.  Given this,
there is no way that this TPL builder worked correctly when just building
shared libs.

I also removed worthless code that was doing FIND_LIBRARY( blas ) and
FIND_LIBRARY( lapack ) right after the EXTERNALPROJECT_ADD() for LAPACK.
These libraries are not built and installed until build time so why in the
hell would you be looking for them at configure time?