Remove *.a static libs in a shared lib build of ZLIB, BOOST, and QT (PHI Kanban #3217 and #232)
The shared lib builds of ZLIB, BOOST, and QT were installing static *.a libs.
That not only messy but sets up the possibility that some software might
choose the *.a libs instead of the *.so libs.  That is not a problem for CMake
(since FIND_LIBRARY() finds *.so files over *.a files by default) but could be
a problem for other build systems.