Fix hypre pointing to blas and lapack, start refactor to allow rebuild of TPL (PHI Kanban #2317 and #3232)
The main thing I did here was to fix how LAPACK and BLAS were being pointed to
in the HYPRE configure.  What was happening is that it was not fining the
special-built BLAS and LAPACK and was finding whatever it could on the system.
This was not caught until I tried to build on james007 where the reference
BLAS does not exist in /usr/lib64/.

The other thing I did here was to start the refactoring that would allow one
to rebuild a downstream TPL given already-installed upstream TPLs.  Before I
can finish this I need to create some macros to remove duplication then I need
to finish up the logic.  These scripts will be much better after that.